+++ optics/lateralFWHM.m
@@ -1,52 +1,42 @@
-function fwhm = aFWHM(wl, NA, n, microscopeType)
-    % AFWHM
+function fwhm = lateralFWHM(wl, NA, microscopeType)
+    % LATERALFWHM
     %
     % Description:
-    %   Theoretical axial resolution 
+    %   Theoretical lateral resolution 
     %
     % Syntax:
-    %   fwhm = aFWHM(wl, n, NA, microscopeType)
+    %   fwhm = lateralFWHM(wl, NA, microscopeType)
     %
     % Inputs:
     %   wl                  float
     %       Wavelength
     %   NA                  float
     %       Numerical aperture (default = Gray 2008, 0.2335)
-    %   n                   float
-    %       Refractive index (default = lens, 1.42)
     %   microscopeType      {'conv', 'conf'}
     %       Confocal or conventional microscope? (default = 'conv')
     %
-    % Notes:
-    %   Axial resolution is ~1.4x better for confocal 
-    %
     % Reference:
-    %   Eq16 in Wilson (2011) Journal of Microscopy, 244, 113-121
+    %   Eq15 in Wilson (2011) Journal of Microscopy, 244, 113-121
     %   NA from Gray et al (2008) IOVS, 49, 467-473
-    %   n from Atchinson ch 16 in Handbook of Visual Optics
     %
     % History:
     %   07Feb2021 - SSP
     %   18Feb2021 - SSP - Added assumptions call rather than hard-coding
     % ---------------------------------------------------------------------
 
-    if nargin < 4
+    if nargin < 3
         microscopeType = 'conv';
     end
 
-    if nargin < 3 || isempty(n)
-        n = assumptions('refractive index'); 
-    end
-
     if nargin < 2 || isempty(NA)
-        NA = assumptions('numerical aperture'); 
-    end 
+        NA = assumptions('numerical aperture');  
+    end
 
     switch microscopeType 
         case 'conv'
-            fwhm = 0.89 * (wl / (n - sqrt(n^2 - NA^2)));
+            fwhm = 0.51 * (wl / NA);
         case 'conf'
-            fwhm = 0.64 * (wl / (n - sqrt(n^2 - NA^2)));
+            fwhm = 0.37 * (wl / NA);
         otherwise
             error('Unrecognized microscope type')
     end